Change HideService name building block to md5 instead of base64
Unique names created by HideService objects included the complete
original file path encoded in base64. The final name depended on the
length of the original path, and could exceed the 256 limit (even more
so when smartcp would add an extra temporary suffix to it).
We now use a fixed-length md5 digest of the name, to ascertain unicity.
Services that used the HideService have an associated json file in which
the original path has been added explicitly.