Add a level with the class Storage above Cache
Move of the methods and arguments of footprints that could be reused in the Storage class.
Change of the use of options to have an interface that could be common with the one that will be used for the Archives.

(cherry picked from commit 8adaace)

(cherry picked from commit 68c99dc)