When dealing with FootprintCopier metaclasses: use six.with_metaclass !
six.add_metaclass does not work properly since it is decorator based.
Because of that, the class is created once (with any existing metaclass)
and only then the new metaclass is called. In most cases, it is harmless
but with footprints based classes, it causes the internal footprints
observer to be filled twice which causes troubles in subsequent
footprints resolutions.