Add new geometries for new dev chain including mid elevation massifs
Since the geometry of the simulation changed (new massifs added), the
Vortex geometry object should also be modified.
New geometry names are the old ones followed by the number of massifs
and an indication on the geometry type (flat terrain only or else with 2 slope values).
A "slope" boolean attribute has been added to these geometries to get
the information elsewhere than in the geometry name.