DBUG after change of PEARP lead time
Feeding the S2M chain with the 2022 DBLE PNT chain implies to use the 0h
PEARP lead time instead of the 18h lead time of D-1 (forecasts now only
available up to 102h instead of 108h) to cover 4 days of forecast (up to
D+4 at 6:00).
As a result the 2 guess files available at 6:00 (D) are :
- a 24h forecast from 6:00 (D-1) --> named Pyymmddhh_24
- a 6h forecast from 0h (D) --> named Pyymmddhh_6

The old algorithm first looked for Pyymmddhh_6 in any situation, but
when dealing with the pseudo forecast from 6:00 (D-1) to 6:00 (D) when
want to use the 24h forecast to have a precipitation cumulation.
The new algorithm now identifies this situation with a 'cumul' boolean
that is False by default by set tu True as soon as a 6h forecast is
missing. When cumul is True, the algorithm stops looking for 6h
forecasts.