DBUG : Do not produce incomplete forcing files
To avoid to backup incomplete forcing files the postprocess method of
sytist workers checks for ExecutionError in "rdict". But in case guess
files are missing, the check_mandatory_resources method rises a InputCheckerError
and prevents any execution launch so that no ExecutionError is raised.
The postprocess method now considers an InputCheckerError as a valid reason to
remove incomplete forcing files.