load_template takes path to template as argument
Previously the first part of the util.config.load_template function
was all about figuring which template file to read, in case the
'tplfile' was passed as '@myfile'. Vortex first looked into the user's
'.vortexrc' file. There was also a lot of code in order to deal with
template files with a version number in their name (removed in an
earlier commit).

This drastically simplifies the template loading capacilities by
letting the callers of 'load_template' figure out the path to the
template file. It effectively removed to ability to override templates
with user's template. I'm not sure this is actually used. If it turns
out this is required, it will be added back.