Reduce TunableAction to tune method
Arguments passed to an action's execute method are used to instanciate
the underlying service object. Some actions inherit from
TunableAction. TunableAction subclasses can add extra parameters to
the service instanciation through the config file or the tune
method. If extra parameters are read from the config file, the
configuration section is usually the value of the 'kind'
attribute. See 'TunableAction._conf_items'.

So far I can't find any example of an action that is configured
through the config file. I'm removing 'TunableAction._conf_items' and
helper methods for now.

Looking at the oper jobs, the only use of tune is to set the jname and
driver parameters. Both are only relevant for actions that instanciate
a Report service (which ones?)