Don't assume structure of namespace for VortexArchiveStore
This is the same as https://github.com/UMR-CNRM/vortex/pull/13 but
for the vortex archive store instead of the cache store.

Currently the `VortexArchiveStore.alternate_netloc` method assumes
that the archive store namespaces will match the following pattern

```
(?P<base>vortex.*)\.archive\.(?P<country>\w+)
```

This removes any constraint on the namespace value. This enables
namespaces based on the `olive-exp` prefix. See
[vortex-olive!3](http://gitlab.meteo.fr/cnrm-gmap/vortex-olive/-/merge_requests/3).

In addition, the namespace suffix is always `.fr`.